refactor: #694 openai/client を責務境界で分割する
